타이머시간 독립적으로 컴포넌트 빼서 분리 & 프로필에서 상호작용 해도 이제 타이머 시간이 초기화되지 않음
